shift card actions to forms
